utilizando heranca para eliminar repeticoes
