+++ Aula-02/exercicio-aula-02/src/com/bclaud/filme/model/Diretor.java
@@ -1,32 +1,15 @@
 package com.bclaud.filme.model;
 
-import com.bclaud.filme.model.enuns.Genero;
+import com.bclaud.filme.enuns.Genero;
 
-public class Diretor {
-    private String nome;
-    private Integer idade;
+public class Diretor extends Funcionario {
     private Integer qtdFilmesDirigidos;
-    private Genero generoEnum;
 
     public Diretor(String nome, Integer idade, Integer qtdFilmesDirigidos, Genero generoEnum) {
-        this.nome = nome;
-        this.idade = idade;
+        super(nome, idade, generoEnum);
         this.qtdFilmesDirigidos = qtdFilmesDirigidos;
-        this.generoEnum = generoEnum;
     }
 
-    public String getNome() {
-        return nome;
-    }
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-    public Integer getIdade() {
-        return idade;
-    }
-    public void setIdade(Integer idade) {
-        this.idade = idade;
-    }
     public Integer getQtdFilmesDirigidos() {
         return qtdFilmesDirigidos;
     }
@@ -34,13 +17,4 @@ public class Diretor {
         this.qtdFilmesDirigidos = qtdFilmesDirigidos;
     }
 
-    public void imprimirInformacoes(){
-        System.out.println(this);
-    }
-
-    @Override
-    public String toString() {
-        return "Diretor [generoEnum=" + generoEnum.getDescricao() + ", idade=" + idade + ", nome=" + nome + "]";
-    }
-    
 }
